code/requirements.txt
absl-py==0.9.0
adversarial-robustness-toolbox==1.3.2
alembic==1.4.0
AllPairs==2.0.1
allpairspy==2.5.0
argon2-cffi==20.1.0
astor==0.8.0
astunparse==1.6.3
async-generator==1.10
attrs==19.3.0
azure-core==1.7.0
azure-storage-blob==12.3.2
backcall==0.2.0
bleach==3.1.5
blinker==1.4
blis==0.7.4
brotlipy==0.7.0
cachetools==4.1.0
catalogue==1.0.0
certifi==2021.5.30
cffi==1.14.0
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
cma==3.0.3
colorama==0.4.4
cycler==0.10.0
cymem==2.0.5
databricks-cli==0.11.0
decorator==4.4.2
defusedxml==0.7.1
dm-tree==0.1.6
docker==4.3.1
docx2txt==0.8
dominate==2.5.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.3
ffmpeg-python==0.2.0
Flask==1.1.2
Flask-Cors==3.0.9
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gevent==1.4.0
gitdb==4.0.5
GitPython==3.1.7
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gorilla==0.3.0
greenlet==1.0.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==4.5.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.1
kiwisolver==1.2.0
lime==0.2.0.1
llvmlite==0.33.0
lxml==4.5.2
Mako==1.1.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.0
matplotlib-inline==0.1.2
mistune==0.8.4
mlflow==1.15.0
msrest==0.6.17
murmurhash==1.0.5
mypy==0.782
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5
nltk==3.5
notebook==6.4.0
numba==0.50.1
numpy==1.20.3
opencv-python==4.4.0.42
opt-einsum==3.1.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.3
parso==0.8.2
patsy==0.5.1
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
preshed==3.0.5
prometheus-client==0.8.0
prometheus-flask-exporter==0.16.0
prompt-toolkit==3.0.18
protobuf==3.13.0
pyasn1==0.4.8
pycparser==2.20
pydub==0.24.1
Pygments==2.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-docx==0.8.11
python-editor==1.0.4
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
pywinpty==1.1.1
PyYAML==5.3.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
querystring-parser==1.2.4
regex==2020.10.28
requests==2.25.1
resampy==0.2.2
scikit-image==0.17.2
scikit-learn==0.22.1
scipy==1.6.3
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.4
spacy==2.3.7
spacy-annotator==2.1.0
SQLAlchemy==1.3.13
sqlparse==0.3.1
srsly==1.0.5
statsmodels==0.11.1
tabulate==0.8.7
tensorboard==2.1.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.10.0
testpath==0.5.0
thinc==7.4.5
tifffile==2020.8.25
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
transformers==3.1.0
typed-ast==1.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
waitress==1.4.4
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
